renaming ResourceConstructor to ResourceOperator
