+++ pennylane/labs/resource_estimation/resource_container.py
@@ -16,7 +16,7 @@ import copy
 from collections import defaultdict
 from dataclasses import dataclass, field
 
-import pennylane.labs.resource_estimation.resource_constructor as rc
+import pennylane.labs.resource_estimation.resource_operator as rc
 
 
 class CompressedResourceOp:
@@ -76,8 +76,8 @@ class CompressedResourceOp:
             >>> print(op_tp)
             QSVT(num_wires=5, num_angles=100)
         """
-        if not issubclass(op_type, rc.ResourceConstructor):
-            raise TypeError("op_type must be of type ResourceConstructor.")
+        if not issubclass(op_type, rc.ResourceOperator):
+            raise TypeError(f"op_type must be a subclass of ResourceOperator. Got {op_type}.")
 
         self._name = op_type.__name__
         self.op_type = op_type
